enumerator: wrap :using helper in :nocov: because it's there for completeness
Since we currently change the underlying enumerator each time
parameters vary, we don't use the parameter helpers directly, only
indirectly one of them when we run initialize.

So the `using` helper is mostly there for completeness, in case some
other day we do things differently. Anyway, currently it means it
should not be considered for coverage.