proxy: remove default enum so that we always build enumerators
This has been done because we don't want to provide an inconsistent
interface in which sometimes you get a cached enumerator instance
and some other times you get a new one. By doing this we lose on
performance for the default case, but we guarantee this consistency
and don't force you to write thread safe code within the iterator
method itself.

However, it would be nice to provide a mechanism to cache enumerator
objects on demand. This might be interesting for a new major
version.

My idea would be to provide a hook for the caller to set up an
external caching mechanism. For example:

caching.store = lambda {}; caching.retrieve = lambda {}

with the appropriate parameters passed in (object, method, and full
parameter list).